Do not remove workdir or move on to distill if early stop was requested.
Note that early stop here does not include timeout.

Early stop requests are caused by errors which need to be shown in the final results. If the workdir is removed and the execution attempt later got preempted, the next attempt will skip this test, as if it previously succeeded. By keeping the workdir, the next attempt would still try to run on the test and either hit the previous issue or avoid it and perform proper fuzzing. Either way is better than skipping the test. The only down side is that left workdir may increase the storage usage. But it will be removed in the next execution.

PiperOrigin-RevId: 767710125