Make Centipede calls aware of termination.
This is to make the termination cleaner. Otherwise when a fuzz test binary terminates the underlying Centipede process could still be running (for a short period, until it gets SIGHUP), sometimes causing flaky failures in e2e tests.

This is done by adding a stop callback (and renaming) to RunCommandWithCallbacks to check for the termination. Also handle the signals to request termination.

PiperOrigin-RevId: 770728386