Allow fuzz test users to define source-code printers
When a fuzz test fails a (best-effort) reproduction GTest is created. For basic domains this works well enough but for complicated user-supplied types built using mappers or other techniques the test will often simply say it operates on an '<unprintable value>' or something equally unhelpful.

This makes it so that if there is a defined `void FuzzTestPrintSourceCode(const T&, std::ostream*)` function in the c++ name search path this function will be used to provide the source-code listing in preference to any other possibility.

PiperOrigin-RevId: 787275306