Add fuzztest::unstable::ParseReproducerValue function.
This function takes a reproducer byte-buffer generated by fuzztest when run with `FUZZTEST_REPRODUCERS_OUT_DIR=` environment variable set and a set of domains and recreates the values which those domains produced when the test failed that generated that reproducer.

This can be used to leverage more advanced analysis and minimization tools than can be reasonably used within the fuzztest itself.

This function is unstable and not part of the stable public API. It may be changed or removed in the future.

PiperOrigin-RevId: 789495655