Abort on external failures in runner mode instead of keep running.
This mainly affects the unit-test mode behavior when dealing with gtest assertion failures.

This is to prepare for persistent mode, where multiple batches run in a single runner process (inside RunnerMain()).
Aborting after failure input ensures Centipede know which batch contains the failure and triage accordingly.

As a side-effect, change a unit-mode e2e test to switch to a non-failing test because the previous test generates an external failure (gTest EXPECT failure) which would cause the runner to abort without destructing the global environment.

PiperOrigin-RevId: 790977540