Use the stop time to set batch execution deadlines.
Do not report crashes on stop conditions.

This allows the stop time to be enforced more precisely, and avoids reporting the potential crashes due to early stopping.

After this we don't need to set batch timeout to the test time limit - it was a quick workaround for the same purpose.

Change the e2e unit test time to 10s to reduce flakiness.

PiperOrigin-RevId: 791875139