Use the stop time to cap command deadlines when running batches for fuzzing.
Do not report crashes on stop conditions.

Exit early in batch condition when termination is requested.

Prolong the unit testing duration to 10s to reduce flakiness.

This allows the stop time to be enforced more precisely instead of having to wait for the entire batch, without reporting the potential crashes due to interrupting the command.

After this we don't need to set batch timeout to the test time limit - it was a quick workaround for the same requirement.

PiperOrigin-RevId: 791875139