Implement persistent mode in runner and `CentipedeCallbacks`.
When persistent mode is enabled, `CentipedeCallbacks` will create a persistent mode socket/connection for each command.
For each batch, CentipedeCallback sends out a request to the runner through the connection.
The runner would in turn process and write the response to the connection without exiting and wait for the next request.

When crash occurs, the command execution would end, which should close the connection on the runner side.
Such event would be detected by `poll()` by `CentipedeCallbacks` to be handled accordingly with `Command::Wait()`.

At the end of the fuzzing loop, `CentipedeCallbacks` sends out a special request to let the runner quit the persistent mode loop and return from `RunnerMain()`.
The special request can be also used to implement finer strategy to restart the persistent mode once a few batches to avoid accumuating too much statefulness to cause unreproducible crashes.

Had to remove a log expectation in the e2e test, since in persistent mode the input may be skipped in the previous batches, which is not printed in the crash log.

PiperOrigin-RevId: 794095134