Mutate newly added fields to protobuf because they can have empty value (e.g., nullopt).
The underlying issue is that, conceptually, the protobuf domain has 2 uninitialized states for each field.

(a) The ProtobufDomain maintains a `map<field_num, field_value>`. If a field is not contained within this map it is considered uninitialized and is unset when converted to a `Message`

(b) fields are represented with `std::optional<field_type>`/`std::vector<field_type>` domain. If the field is empty (e.g., `std::nullopt`) then, similarly, the field is considered uninitialized and is unset when converted to a `Message`.

`Mutate` ends up doing nothing from the perspective of the final output `Message` as we transition between these 2 states. Thus, when adding a field, first we initialize and then perform a mutation to make sure the field is not empty with a high probability.

PiperOrigin-RevId: 795468043