Report crash summary when replaying a corpus or a crashing input.
At this point, the crash summary is still reported only when we have a corpus
database and when we're running `CentipedeMain`. That means that the crash
summary won't be reported in single-process mode or for one-off runs without
the corpus database. Enabling this will require a more involved reorganization
of the code: we need the concepts of crash metadata (category/type, signature,
description) not to be entangled with the Centipede's fuzzing loop.

PiperOrigin-RevId: 799670318