Clean up persistent mode when only loading shards.
The function `Centipede::FuzzingLoop()` has an early return when
`env_.load_shard_only` is true, but in this case it wasn't cleaning up the
persistent mode. This was resulting in unexpected behavior, e.g., binaries
instrumented for coverage collection would sometimes be terminated before
the coverage data had a chance of being collected.

PiperOrigin-RevId: 815984240