Introduce a new crash deduplication mechanism in Centipede.
The change is guarded by `FUZZTEST_INTERNAL_BETTER_CRASH_DEDUPLICATION`.

As part of this change, we also need to hash the failure description as the
crash signature when a dispatcher is not used. This is because the new crash
deduplication expects that the crash signature can be used as part of a file
name.

In case the dispatcher is used, it'll have to ensure this requirement.

PiperOrigin-RevId: 828969421