Handle unexpected termination in Centipede's crash handling.
If the test binary exits without dumping the failure description, we assign
"unexpected-termination" as the fallback failure description. In the most common
case, this will handle exits via `std::_Exit` (immediate exits without cleanup),
which underly logging via QCHECK / QFATAL.

One corner case is an exit via `std::_Exit(0)`, which appears as a successful
execution, but should technically be considered a bug in the fuzz test. We
won't bother handling this for now, since it should be sufficiently rare, and
distinguishing this from successful execution of non-FuzzTest targets would be
cumbersome.

PiperOrigin-RevId: 840749159